backend/requirements.txt
# Core FastAPI Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ML & Data Processing
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# NFL Data
nfl_data_py>=0.3.0

# WebSocket & Real-Time
websockets>=12.0
python-dotenv>=1.0.0

# Future: Vision Agents (uncomment when ready for real video processing)
# vision-agents>=0.1.0
# ultralytics>=8.0.0
# stream-python>=4.0.0
